requirements.txt
matplotlib==3.7.2
plotly==5.15.0
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.5
colorama==0.4.6
tabulate==0.9.0
